+++ src/lib/index.js
@@ -1,3 +1,8 @@
-import RangeSlider from './components/RangeSlider'
+import React, { forwardRef } from 'react'
+import RangeSliderInner from './components/RangeSlider'
+
+const RangeSlider = forwardRef((props, ref) => {
+  return <RangeSliderInner {...props} forwardedRef={ref} />
+})
 
 export default RangeSlider